+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - `RealTensor` typing class.
 - Ratios conversions functions for masking.
 - `invalid_shape` argument to `get_shape` function.
+- `return_default_on_invalid` option to `get_shape` and `get_ndim` fns.
 
 ### Modified
 - `move_to` function now supports dataclasses.